Calculate ensemble power spectra

Calculate ensemble power spectra using 4 spectral methods:

Lomb-Scargle periodogram

REDFIT (Mudelsee et al, 2009)

Multitaper Method (Thomson et al. 1982)

nuspectral (Mathias et al 2004) (no confidence estimation)

For REDFIT and MTM, only the 95 For multiple ensemble members, the median of the ensemble of 95 For Lomb-Scargle, the parameter `probs` determines the quantiles of the surrogate spectrum distribution extracted as confidence limits.

computeSpectraEns(
  time,
  values,
  max.ens = NA,
  method = "mtm",
  probs = 0.95,
  gaussianize = TRUE,
  ofac = 4,
  padfac = 2,
  tbw = 3,
  wgtrad = 1,
  sigma = 0.02,
  mtm_null = "power_law"
)

Arguments

time
LiPD "variable list" or vector of year/age values

values
LiPD "variable list" or vector of values

max.ens
Maximum number of ensemble members to analyze

probs
vector of probabilities for the siginificance levels. To avoid such computations, pass `probs = NA`.

gaussianize
Boolean flag indicating whether the values should be mapped to a standard Gaussian prior to analysis.

ofac
oversampling factor for lomb::lsp and dplR::redfit

padfac
padding factor for astrochron::mtm___ functions

tbw
time-bandwidth product for astrochron::mtm___ functions

wgtrad
radius of the nuspectral::nuwaveletcoeff weight function (non-dimensional units)

sigma
decay parameter of the nuspectral::nuwaveletcoeff wavelets

mtm_mull
null hypothesis for the detectiion of significant peaks in `astrochron::mtm()`. Possible choices: AR(1), power_law, or ML96.

Value

a list of ensemble spectra results

freqs: vector of frequencies

power: vector of spectral powers

power.CL: matrix of confidence limits for spectral power

Details

Calculate ensemble power spectra

References

Mudelsee, M., D. Scholz, R. Röthlisberger, D. Fleitmann, A. Mangini, and E. W. Wolff (2009), Climate spectrum estimation in the presence of timescale errors, Nonlinear Processes in Geophysics, 16(1), 43–56, doi:10.5194/npg-16-43-2009.

Mathias, A., F. Grond, R. Guardans, D. Seese, M. Canela, and H. Diebner (2004), Algorithms for spectral analysis of irregularly sampled time series, Journal of Statistical Software, Articles, 11(2), 1–27, doi:10.18637/jss.v011.i02.

Thomson, D. J. (1982), Spectrum estimation and harmonic analysis, Proc. IEEE, 70(9), 1055–1096.

See also