Create the oxcal model script

Create the oxcal model script

createOxcalModel(
  cdf,
  depths.to.model = NA,
  depth.interval = 10,
  thin = 50,
  n.it = 10000,
  events.per.unit.length = 1,
  events.per.unit.length.uncertainty = NA,
  seq.type = "P_Sequence",
  top.boundary = "Boundary",
  bottom.boundary = "Boundary",
  outlier.prob = 0.05,
  cal.curve = "intcal13"
)

Arguments

cdf
A data.frame of chron measurement data (output from createChronMeasInputDf())

depths.to.model
A vector of depths to model (default = create a sequence using depth.interval)

depth.interval
Depth interval to use to create depths.to.model (default = 10)

thin
Thinning parameter for OxCal (default = 50)

n.it
Number of iteration in model (default = 10000)

events.per.unit.length
How many events per unit length for the Poisson model? This controls the flexibility on the model, the lower the number the more flexibile.

events.per.unit.length.uncertainty
Uncertainty on events.per.unit.length in orders of magnitude. This makes for better models but *greatly* increases the run time. If NA or < 0 it runs without uncertainty (default = NA)

seq.type
What type of Oxcal sequence to use? (default = "P_Sequence")

top.boundary
What type of Oxcal boundary to use for the top (default = "Boundary")

bottom.boundary
What type of Oxcal boundary to use for the bottom (default = "Boundary")

outlier.prob
What's the probability of a date being an outlier if not otherwise specified (default= 0.05)

cal.curve
What calibration curve to use? (default = "intcal13")

Value

See also