Map an ageEnsemble variable from a chron model to a paleoMeasurement Table

Copies an ageEnsemble from chronData (model) to paleoData (measurementTable), by matching depth and interpolating (extrapolating) as necessary.

mapAgeEnsembleToPaleoData(
  L,
  age.var = "ageEnsemble",
  chron.depth.var = "depth",
  paleo.depth.var = "depth",
  paleo.age.var = "age",
  map.median = NA,
  paleo.num = NA,
  paleo.meas.table.num = NA,
  chron.num = NA,
  model.num = NA,
  ens.table.num = 1,
  max.ens = NA,
  paleo.depth.range = NA,
  strict.search = FALSE
)

Arguments

L

A LiPD object - an R serialization of a single LiPD file. It's a list, and is typically created by `readLipd()`

age.var

name of the age ensemble variable to search for

chron.depth.var

name of the depth variable to search for in the ensemble table

paleo.depth.var

name of the depth variable to search for in the paleo measurement table

paleo.age.var

name of the age variable to search for in the paleo measurement table, critical if no depth data are available

map.median

Do you want to also map the median of the ageEnsemble to paleoData? (TRUE or FALSE, or NA, the default, which will do so unless that variable already exists)

paleo.num

an integer that corresponds to paleo.numData object (L$paleoData[[?]]) has the measurementTable you want to modify

paleo.meas.table.num

an integer that corresponds to paleo.num measurementTable you want to add the ensemble to?

chron.num

an integer that corresponds to chron.numData object (L$chronData[[?]]) has the model you want to get the ensemble from

model.num

an integer that corresponds to chron.num model you want to get the ensemble from?

ens.table.num

an integer that corresponds to ensembleTable you want to get the variable from?

max.ens

Maximum number of ensemble members to map

paleo.depth.range

A two element vector eg: c(min,max), that describes the range of in the paleo data depth over which to map the data. Depths outside this range will correspond to NAs in the age ensemble. By default (default = NA) this is the full range of the paleo.depth.var.

strict.search

Use a strict.search to look for the ageEnsemble and depth variables. TRUE(default) or FALSE.

Value

L a lipd object

See also