Map a TS object

Create a stamen or line map of the location of a list of LiPD objects

mapTs(
  TS,
  color = "archiveType",
  size = 6,
  shape = 16,
  lat.range = NA,
  lon.range = NA,
  ...
)

Arguments

TS
A list of LiPD objects

color
variable (in TS) by which to color, or a string specifying a static ggplot color

size
Size of the location marker

shape
variable (in TS) by which to adjust shape, or an integer specifying a static ggplot shape

lat.range
Latitudes to use to create baseMap range (default = NA, which determines from TS)

lon.range
Longitudes to use to create baseMap range (default = NA, which determines from TS)

...
Arguments passed on to baseMap

lat

latitude(s) range to map

lon

longitude(s) range to map

map.type

"stamen" or "line"

f

buffer for the map range

restrict.map.range

TRUE or FALSE. Trim the size of the map to the points, for "line" map type

bound.circ

Draw a boundary circle around a polar projection. TRUE or FALSE(default).

global

Should the scope of the map be global? TRUE or FALSE(default).

projection

Map project. All options on: ?mapproject

extend.range

increase the span of the map by this much (lat/long degrees)

Value

ggmap object

Long-form example

View a full-fledged example of how to use this function.

See also

Author

Nick McKay