Compare chron ensemble with paleoData age-model

Plots the difference of an chron ensembleTable with the paleoData age

plotChronEnsDiff(
  L,
  ageEnsVar = "ageEnsemble",
  age.var = "age",
  depth.var = "depth",
  paleo.num = NA,
  paleo.meas.table.num = NA,
  chron.num = NA,
  model.num = NA,
  ens.table.num = NA,
  max.ensemble.members = NA,
  strict.search = FALSE,
  probs = c(0.025, 0.25, 0.5, 0.75, 0.975),
  x.bin = NA,
  y.bin = NA,
  n.bins = 100,
  color.low = "white",
  color.high = "grey70",
  alp = 1,
  color.line = "Black",
  line.width = 1,
  add.to.plot = ggplot2::ggplot(),
  n.ens.plot = 5,
  color.ens.line = "red",
  alp.ens.line = 0.7
)

Arguments

L
A LiPD object - an R serialization of a single LiPD file. It's a list, and is typically created by `readLipd()`

ageEnsVar
name of the age ensemble variable in the chronData to search for

age.var
name of the age variable in the paleoData to search for

depth.var
name of the depth variable to search for

paleo.num
an integer that corresponds to paleo.numData object (L$paleoData[[?]]) has the measurementTable you want to modify

paleo.meas.table.num
an integer that corresponds to paleo.num measurementTable you want to add the ensemble to?

chron.num
an integer that corresponds to chron.numData object (L$crhonData[[?]]) has the model you want to get the ensemble from

model.num
an integer that corresponds to chron.num model you want to get the ensemble from?

ens.table.num
an integer that corresponds to chron.num model ensembleTable you want to get the ensemble from?

max.ensemble.members
Maximum number of ensemble members to map

strict.search
Use a strict.search to look for the ageEnsemble and depth variables. TRUE(default) or FALSE #' @param probs quantiles to calculate and plot

x.bin
vector of bin edges over which to bin.

y.bin
vector of bin edges over which to bin.

n.bins
number bins over which to calculate intervals. Used to calculate x.bin if not provided.

color.low
Band color of the outer most band.

color.high
Band color of the inner most band.

alp
Transparency of the band plot

color.line
Line color (following ggplot rules)

line.width
Width of the line

add.to.plot
A ggplot object to add this plot to. Default is ggplot() .

n.ens.plot
Number of ensemble members to plot

color.ens.line
color of the ensemble lines

alp.ens.line
transparency of the lines

Value

A ggplot object

See also

Author

Nick McKay