Map ensemble pca loadings and plot PC timeseries

Map ensemble pca loadings and plot PC timeseries

plotPcaEns(
  ens.pc.out,
  TS,
  map.type = "line",
  which.pcs = c(1, 2),
  f = 0.2,
  high.color = "red",
  low.color = "blue",
  dot.size = 5,
  restrict.map.range = TRUE,
  shape.by.archive = TRUE,
  projection = "mollweide",
  bound.circ = TRUE,
  probs = c(0.025, 0.25, 0.5, 0.75, 0.975),
  which.leg = 1,
  legend.position = c(0.5, 0.5),
  color
)

Arguments

ens.pc.out

results of pcaEns()

TS

Timeseries object http://nickmckay.github.io/LiPD-utilities/r/index.html#what-is-a-time-series used in the pcaEns() analysis

map.type

"google" or "line"

which.pcs

vector of PCs to plot. Choose two. c(1,2) is default.

f

zoom buffer for plotting

high.color

color for the high end of the scale

low.color

color for the low end of the scale

dot.size

How big are the dots on the map

restrict.map.range

TRUE or FALSE. Trim the size of the map to the points, for "line" map type

shape.by.archive

TRUE or FALSE. Use archiveType to assign shapes.

projection

Map project. All options on: ?mapproject

bound.circ

For polar projects, draw a boundary circle? TRUE or FALSE

probs

quantiles to calculate and plot in the PC timeseries

which.leg

which map legend to include in the summary plot?

legend.position

Where to put the map legend?

color

deprecated. Use high.color and low.color instead

Value

A gridExtra ggplot object

See also

Author

Nick McKay