Plot ensemble regression results

Creates a suite of plots to characterize the results of an ensemble regression.

plotRegressEns(
  reg.ens,
  alp = 0.2,
  quantiles = c(0.025, 0.5, 0.975),
  font.size = 10
)

Arguments

reg.ens
output of regressEns()

alp
Transparency of the scatter plot.

quantiles
quantiles to calculate and plot

Value

A list of ggplot objects

YPlot - ribbon plot of the prectictand timeseries over the interval of overlap

XPlot - ribbon plot of the predictor timeseries over the interval of overlap

scatterplot - ensemble scatter plot of the predictor and predictand timeseries over the interval of overlap

mHist - distribution of ensemble regression slopes

bHist - distribution of ensemble regression intercepts

modeledYPlot - ribbon plot of values modeled by the ensemble regression, incorporating age uncertainty in both the regression and the predictor timeseries

summaryPlot - grid.arrange object of all the regression plots

See also

Author

Nick McKay