Plot a summary figure

shows a map, timeseries, and age model diagram, and basic simple metadata

plotSummary(
  L,
  paleo.age.var = "age",
  paleo.data.var = NA,
  chron.number = NA,
  paleo.meas.num = NA,
  chron.meas.num = NA,
  chron.depth.var = "depth",
  chron.age.var = "age",
  dotSize = 5,
  summary.font.size = 10,
  text.width = 400/summary.font.size,
  legend.position = c(0.7, 0.3),
  ...
)

Arguments

L
A LiPD object - an R serialization of a single LiPD file. It's a list, and is typically created by `readLipd()`

paleo.age.var
variableName to use for x axis of the paleo plot ("age" by default)

paleo.data.var
variableName to use for the y axis of the paleo plot (NA by default, which lets you choose)

chron.number
chron.numData object to use (NA by default, will ask if needed)

paleo.meas.num
paleo.num measurement table to use (NA by default, will ask if needed)

chron.meas.num
chron.num measurement table to use (NA by default, will ask if needed)

chron.depth.var
variableName to use for chron depth ("depth" by default)

chron.age.var
variableName to use for chron age ("age" by default)

...
Arguments passed on to plotChronEns

depth.var

variableName to use for depth ("depth" by default)

age.var

ariableName to use for age ensemble ("ageEnsemble" by default)

model.num

chron.numData model to use (NA by default, will ask if needed)

probs

quantiles to calculate and plot

n.bins

number bins over which to calculate intervals. Used to calculate x.bin if not provided.

x.bin

vector of bin edges over which to bin.

y.bin

vector of bin edges over which to bin.

color.low

Band color of the outer most band.

color.high

Band color of the inner most band.

alp

Transparency of the band plot

color.line

Line color (following ggplot rules)

line.width

Width of the line

add.to.plot

A ggplot object to add this plot to. Default is ggplot() .

n.ens.plot

Number of ensemble members to plot

color.ens.line

color of the ensemble lines

alp.ens.line

transparency of the lines

dist.color

distribution color (following ggplot rules)

dist.type

"violin" (default), "up" for one-sided distributions pointed up, "down" for one-sided distributions pointed down

dist.thick

thickness of the line around the distribution

truncate.dist

truncate probability density values below this number. NA (default) means no truncation

dist.scale

controls the vertical span of the probability distribution. Approximately the vertical fraction of the plot that the distribution will cover.

add.paleo.age.depth

add a line that shows the paleoData age depth.

paleo.number

paleo.num number for the paleoData age-depth

meas.num

which measurement Table for the paleoData age-depth

color.line.paleo

line color of the paleoData age-depth (following ggplot rules)

dot.size
what size dot for the chron plot? Only used if not plotting by plotChronEns() (default = 5)

Value

A gridArrange of ggplot grobs

See also

Author

Nick McKay

Examples

if (FALSE) {
myPlot = summaryPlot(L)
}