Plot an ensemble timeseries as a set of lines

Plot an ensemble timeseries as a set of lines. Useful for displaying a handful of ensemble members to characterize individual paths.

plotTimeseriesEnsLines(
  add.to.plot = ggplot(),
  X,
  Y,
  alp = 0.2,
  color = "blue",
  n.ens.plot = 100,
  na.rm = TRUE
)

Arguments

add.to.plot
A ggplot object to add these lines to. Default is ggplot() .

X
A LiPD variable list to plot, including values, units, names, and more

Y
A LiPD variable list to plot, including values, units, names, and more

alp
Line transparency

color
Either 1) A line color (following ggplot rules) to use for all lines (e.g., "blue"), 2) An RColorBrewer pallette to repeat over the lines (e.g. "Blues") or 3) a vector specifying the color for all lines (e.g., c("red","white","blue"))

n.ens.plot
Whats the maximum number of lines to plot?

na.rm
Remove NAs from X and Y? Set to FALSE to preserve line breaks where data are missing. (default = TRUE)

Value

A ggplot object

See also

Author

Nick McKay