Proportion of alternate hypotheses

This is an internal function that calculates an estimate of a, the proportion of alternative hypotheses, using one of several methods.

propAlt(
  pvals,
  adjustment.method = "mean",
  adjustment.args = list(edf.lower = 0.8, num.steps = 20)
)

Arguments

pvals
a vector of pvals on which to conduct the multiple testing

adjustment.method
method for increasing the power of the procedure by estimating the proportion of alternative p-values, one of "mean", the modified Storey estimator that we suggest in Ventura et al. (2004), "storey", the method of Storey (2002), or "two-stage", the iterative approach of Benjamini et al. (2001)

adjustment.args
arguments to adjustment.method; see propAlt() for description, but note that for "two-stage", qlevel and fdr.method are taken from the qlevel and method arguments to fdr()

qlevel
the proportion of false positives desired

Value

estimate of a, the number of alternative hypotheses

See also

Author

Chris Paciorek

Examples

a <- propAlt(pvals,adjustment.method="mean")

#> Error in propAlt(pvals, adjustment.method = "mean"): object 'pvals' not found