Generate an oxcal Age Model and add it into a LiPD object

This is a high-level function that uses oxcal to simulate an age model, and stores this as an age-ensemble in a model in chronData. If needed input variables are not entered, and cannot be deduced, it will run in interactive mode. See Bronk Ramsey et al. 2008 doi:10.1016/j.quascirev.2007.01.019

runOxcal(
  L,
  chron.num = NA,
  meas.table.num = NA,
  oxcal.path = "~/OxCal/",
  model.num = NA,
  remove.rejected = TRUE,
  overwrite = TRUE,
  max.ens = 1000,
  static.reservoir.age = NA,
  static.reservoir.age.unc = NA,
  surface.age = NA,
  surface.age.unc = NA,
  surface.age.depth = 0,
  depth.interval = NA,
  events.per.unit.length = NA,
  events.per.unit.length.uncertainty = 0,
  outlier.prob = 0.05,
  cal.curve = "IntCal13",
  ...
)

Arguments

L
A LiPD object - an R serialization of a single LiPD file. It's a list, and is typically created by `readLipd()`

chron.num
the number of the chronData object that you'll be working in

meas.table.num
the number of the measurementTable you'll be working in

model.num
chron.numModel do you want to use?

remove.rejected
don't write out dates that are marked as rejected

overwrite
overwrite files and directories

max.ens
the maximum number of ensembles to load in (default = 1000)

static.reservoir.age
optionally assign a deltaR to all radiocarbon dates

static.reservoir.age.unc
optionally assign a deltaR uncertainty to all radiocarbon dates

...
Arguments passed on to createChronMeasInputDf

lab.id.var

age.14c.var

age.14c.uncertainty.var

age.var

age.uncertainty.var

depth.var

reservoir.age.14c.var

reservoir.age.14c.uncertainty.var

rejected.ages.var

split.ages

if there's an age_type column, and only one age column, intelligently split between age and age14C

Value

L The single LiPD object that was entered, with methods, ensembleTable, summaryTable and distributionTable added to the chronData model.

See also

Author

Nick McKay

Examples

if (FALSE) {
#Run in interactive mode:
L = runOxcal(L)

#Run in noninteractive mode, describing everything:
L = runOxcal(L,chron.num = 1, meas.table.num = 1, model.num = 3, bacon.dir = "~/Bacon/",site.name = "MSB2K", cc = 1)
}