create a scree plot for PCA analysis

create a scree plot for PCA analysis

plotScreeEns(pcaout, null.color = "red", null.significance = 0.05, ...)

Arguments

pcaout
a list of results output from pcaEns()

null.color
color of the line of the null hypothesis results

null.significance
Significance level to of the null to plot (default = 0.05).

...
Arguments passed on to plotTimeseriesEnsRibbons

X

A LiPD variable list to plot, including values, units, names, and more

Y

A LiPD variable list to plot, including values, units, names, and more

probs

a vector of probabilities to plot as ribbons. It will create bands as ribbons of quantiles moving inward. If there's an odd number, it plots the middle quantile as a line.

color.low

Color of the outermost band; the extreme quantiles of the distribution

color.high

Color of the innermost band; the central quantiles of the distribution

color.line

Line color (following ggplot rules)

color.vector

A vector (of length equal to the number of bands) that specifies the colors for the ribbons from the outermost band in (default = NA). Colors specified as string according to ggplot2 conventions. If present, this overrules color.high and color.low

line.width

Width of the line

n.bins

number bins over which to calculate intervals. Used to calculate x.bin if not provided.

x.bin

vector of bin edges over which to bin.

y.bin

vector of bin edges over which to bin.

alp

alpha (transparency) parameter for the ribbons

add.to.plot

A ggplot object to add this plot to. Default is ggplot() .

export.quantiles

If TRUE, teturn the plotted quantiles rather than the plot

Value

a ggplot plot

See also